+++ sdks/python/apache_beam/runners/direct/direct_runner.py
@@ -32,8 +32,6 @@ from apache_beam.runners.runner import PipelineResult
 from apache_beam.runners.runner import PipelineRunner
 from apache_beam.runners.runner import PipelineState
 from apache_beam.runners.runner import PValueCache
-from apache_beam.runners.runner import group_by_key_input_visitor
-from apache_beam.runners.runner import flatten_input_visitor
 from apache_beam.utils.pipeline_options import DirectOptions
 
 
@@ -70,8 +68,6 @@ class DirectRunner(PipelineRunner):
     MetricsEnvironment.set_metrics_supported(True)
     logging.info('Running pipeline with DirectRunner.')
     self.consumer_tracking_visitor = ConsumerTrackingPipelineVisitor()
-    pipeline.visit(group_by_key_input_visitor())
-    pipeline.visit(flatten_input_visitor())
     pipeline.visit(self.consumer_tracking_visitor)
 
     evaluation_context = EvaluationContext(
