+++ sdks/python/apache_beam/runners/runner.py
@@ -94,65 +94,6 @@ def create_runner(runner_name):
             runner_name, ', '.join(_ALL_KNOWN_RUNNERS)))
 
 
-def group_by_key_input_visitor():
-  # Imported here to avoid circular dependencies.
-  from apache_beam.pipeline import PipelineVisitor
-
-  class GroupByKeyInputVisitor(PipelineVisitor):
-    """A visitor that replaces `Any` element type for input `PCollection` of
-    a `GroupByKey` or `GroupByKeyOnly` with a `KV` type.
-
-    TODO(BEAM-115): Once Python SDk is compatible with the new Runner API,
-    we could directly replace the coder instead of mutating the element type.
-    """
-
-    def visit_transform(self, transform_node):
-      # Imported here to avoid circular dependencies.
-      # pylint: disable=wrong-import-order, wrong-import-position
-      from apache_beam import GroupByKey, GroupByKeyOnly
-      from apache_beam import typehints
-      if isinstance(transform_node.transform, (GroupByKey, GroupByKeyOnly)):
-        pcoll = transform_node.inputs[0]
-        input_type = pcoll.element_type
-        # If input_type is not specified, then treat it as `Any`.
-        if not input_type:
-          input_type = typehints.Any
-
-        if not isinstance(input_type, typehints.TupleHint.TupleConstraint):
-          if isinstance(input_type, typehints.AnyTypeConstraint):
-            # `Any` type needs to be replaced with a KV[Any, Any] to
-            # force a KV coder as the main output coder for the pcollection
-            # preceding a GroupByKey.
-            pcoll.element_type = typehints.KV[typehints.Any, typehints.Any]
-          else:
-            # TODO: Handle other valid types,
-            # e.g. Union[KV[str, int], KV[str, float]]
-            raise ValueError(
-                "Input to GroupByKey must be of Tuple or Any type. "
-                "Found %s for %s" % (input_type, pcoll))
-
-  return GroupByKeyInputVisitor()
-
-def flatten_input_visitor():
-  # Imported here to avoid circular dependencies.
-  from apache_beam.pipeline import PipelineVisitor
-
-  class FlattenInputVisitor(PipelineVisitor):
-    """A visitor that replaces the element type for input ``PCollections``s of a
-     ``Flatten`` transform with that of the output ``PCollection``.
-    """
-
-    def visit_transform(self, transform_node):
-      # Imported here to avoid circular dependencies.
-      # pylint: disable=wrong-import-order, wrong-import-position
-      from apache_beam import Flatten
-      if isinstance(transform_node.transform, Flatten):
-        output_pcoll = transform_node.outputs[None]
-        for input_pcoll in transform_node.inputs:
-          input_pcoll.element_type = output_pcoll.element_type
-
-  return FlattenInputVisitor()
-
 class PipelineRunner(object):
   """A runner of a pipeline object.
 
@@ -186,16 +127,6 @@ class PipelineRunner(object):
           logging.error('Error while visiting %s', transform_node.full_label)
           raise
 
-    # Replace the element type for inputs to a GroupByKey operation with a
-    # KV type hint that would allow runners to choose a standard coder
-    # (TupleCoder) for these elements.
-    pipeline.visit(group_by_key_input_visitor())
-
-    # Replace the element type of inputs to a Flatten operation with that of the
-    # output of Flatten.
-    # TODO: This is only needed for the Dataflow runner, fix once the new runner
-    # API is adopted.
-    pipeline.visit(flatten_input_visitor())
     pipeline.visit(RunVisitor(self))
 
   def clear(self, pipeline, node=None):
